Move plugin tests
