+++ conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = "tests.fixtures"
+pytest_plugins = ["tests.fixtures", "saleor.plugins.tests.fixtures"]
