+++ saleor/plugins/avatax/tests/cassettes/test_calculate_checkout_shipping.yaml
