+++ saleor/plugins/avatax/tests/cassettes/test_calculate_checkout_total[False-40.00-49.20-0.0-False].yaml
