+++ saleor/plugins/avatax/tests/cassettes/test_calculate_checkout_total[True-20.33-25.00-0.0-True].yaml
