+++ saleor/plugins/tests/test_checks.py
